Moved dummy history data to tests/
